Update index.html and style.css
- In index.html, more-box div has a width slightly more than the intro-box. Applying
border to the whole section wasn't an option since the width of more-box
is less than the section width which has the width of the body. So,
more-box is now wrapped with a div that has a width of the entire body.
This way we can apply border to the entire section.

- In style.css, added CSS color variables for repeatedly used colors and
also applied styles to the more-box section of the site.

- The skill bars aren't filled yet and will be done using JS in a future
commit